[SDR-200] Bug fix RequeryResultConverter * Add RoleRepository Test cases

+++ spring-data-requery-kotlin/src/main/kotlin/org/springframework/data/requery/kotlin/repository/query/RequeryResultConverter.kt
@@ -17,7 +17,7 @@ object RequeryResultConverter {
             when(result) {
                 is Tuple ->
                     when {
-                        result.count() == 1 -> result.get(0)
+                        result.count() == 1 -> result.get(0) as Int
                         else -> result
                     }
                 else -> result
